Stabilised and optimised client-side OrbitDB connection

+++ backend/orbit.js
@@ -1,54 +1,60 @@
-const config = require('../config')
 const fs = require('fs');
-const IPFS = require('ipfs')
-const OrbitDB = require('orbit-db')
 
-const ipfsOptions = {
-  start: true,
-  EXPERIMENTAL: {
-    pubsub: true,
-  },
-  config: {
-    Addresses: {
-      Swarm: [
-        '/dns4/ws-star.discovery.libp2p.io/tcp/443/wss/p2p-websocket-star',
-      ]
+if (!fs.existsSync('data.json')) {
+
+  const IPFS = require('ipfs')
+  const OrbitDB = require('orbit-db')
+
+  const ipfsOptions = {
+    start: true,
+    EXPERIMENTAL: {
+      pubsub: true,
     },
-  }
-}
-
-const ipfs = new IPFS(ipfsOptions)
-
-ipfs.on('error', (e) => console.error(e))
-
-ipfs.on('ready', async () => {
-
-  const orbitdb = new OrbitDB(ipfs)
-
-  const db = await orbitdb.open(config.address, { sync: true })
-  await db.load()
-
-  db.events.on('replicated', () => {
-
-    let data = {
-      "x": getArray(db, 'x'),
-      "y": getArray(db, 'y'),
-      "truth": getArray(db, 't')
+    config: {
+      Addresses: {
+        Swarm: [
+          '/dns4/ws-star.discovery.libp2p.io/tcp/443/wss/p2p-websocket-star',
+        ]
+      },
     }
-    fs.writeFileSync('data.json', JSON.stringify(data));
+  }
 
-    // Disconnection TODO: stabilise
-    /*
-    orbitdb.disconnect()
-    ipfs.stop(() => {})
-    */
+  const ipfs = new IPFS(ipfsOptions)
+
+  ipfs.on('error', (e) => console.error(e))
+
+  ipfs.on('ready', async () => {
+
+    const orbitdb = new OrbitDB(ipfs)
+
+    const db = await orbitdb.open('/orbitdb/QmcHKBHmXmx3HHi1opqAVqF4E3qwwkmJPQ8CeVX4ybc4xJ/trainingdata')
+    await db.load()
+
+    db.events.on('replicated', async () => {
+      try {
+        let data = {
+          "x": getArray(db, 'x'),
+          "y": getArray(db, 'y'),
+          "truth": getArray(db, 't')
+        }
+        fs.writeFileSync('data.json', JSON.stringify(data));
+        await orbitdb.disconnect()
+        await ipfs.stop(() => {})
+        process.exit()
+      }
+      catch(error) {
+        // OrbitDB considers itself loaded before it is readable, catch this
+        console.log('Loading OrbitDB...')
+      }
+
+    })
 
   })
+  function getArray(db, name) {
 
-})
+    let entry = db.get(name)
+    return entry[0].array
 
-function getArray(db, name) {
-
-  return db.get(name)[0].array
+  }
 
 }
